packagisation du projet partie1/2
